Move high-level file function to pingrid
