+++ fbfmaproom/fbfmaproom.py
@@ -7,7 +7,6 @@ import io
 from functools import lru_cache
 import datetime
 import urllib.parse
-import yaml
 import json
 import numpy as np
 import pandas as pd
@@ -901,36 +900,6 @@ def borders(pathname, mode):
 # Endpoints
 
 
-def image_resp(im):
-    cv2_imencode_success, buffer = cv2.imencode(".png", im)
-    assert cv2_imencode_success
-    io_buf = io.BytesIO(buffer)
-    resp = flask.send_file(io_buf, mimetype="image/png")
-    return resp
-
-
-def yaml_resp(data):
-    s = yaml.dump(data, default_flow_style=False, width=120, allow_unicode=True)
-    resp = flask.Response(response=s, mimetype="text/x-yaml")
-    resp.headers["Cache-Control"] = "private, max-age=0, no-cache, no-store"
-    return resp
-
-
-def tile(dae, tx, ty, tz, clipping=None, test_tile=False):
-    z = pingrid.produce_data_tile(dae.data_array, tx, ty, tz)
-    im = (z - dae.min_val) * 255 / (dae.max_val - dae.min_val)
-    im = pingrid.apply_colormap(im, dae.colormap)
-    if clipping is not None:
-        draw_attrs = pingrid.DrawAttrs(
-            BGRA(0, 0, 255, 255), BGRA(0, 0, 0, 0), 1, cv2.LINE_AA
-        )
-        shapes = [(clipping, draw_attrs)]
-        im = pingrid.produce_shape_tile(im, shapes, tx, ty, tz, oper="difference")
-    if test_tile:
-        im = pingrid.produce_test_tile(im, f"{tz}x{tx},{ty}")
-    return image_resp(im)
-
-
 @SERVER.route(
     f"{TILE_PFX}/pnep/<int:tz>/<int:tx>/<int:ty>/<country_key>/<season_id>/<int:target_year>/<int:issue_month_idx>/<int:freq>"
 )
@@ -942,7 +911,7 @@ def pnep_tiles(tz, tx, ty, country_key, season_id, target_year, issue_month_idx,
     dae = select_pnep(country_key, issue_month0, target_month0, target_year, freq)
     p = tuple(CONFIG["countries"][country_key]["marker"])
     clipping, _ = retrieve_geometry(country_key, p, "0", None)
-    resp = tile(dae, tx, ty, tz, clipping)
+    resp = pingrid.tile(dae, tx, ty, tz, clipping)
     return resp
 
 
@@ -981,7 +950,7 @@ def vuln_tiles(tz, tx, ty, country_key, mode, year):
             for _, r in df.iterrows()
         ]
         im = pingrid.produce_shape_tile(im, shapes, tx, ty, tz, oper="intersection")
-    return image_resp(im)
+    return pingrid.image_resp(im)
 
 
 def cache_stats(f):
@@ -1023,7 +992,7 @@ def stats():
         cache_stats=cs,
         process_stats=ps,
     )
-    return yaml_resp(rs)
+    return pingrid.yaml_resp(rs)
 
 
 @SERVER.route(f"{PFX}/pnep_percentile")
